Alertas

+++ src/main/resources/static/js/inventory_see.js
@@ -58,19 +58,34 @@ function mostrarInventario(inventario) {
 			<div style="text-align: center;">
 				<p>${producto.store}</p>
 			</div>
+			<div>
+				<button type="button" id="btn-update-${producto.id}" name=${producto.id}>Actualizar</button>
+                <button type="button" id="btn-cancel-${producto.id}" name=${producto.id}>Eliminar</button>
+			</div>
 		`;
-		const divAcciones = document.createElement("DIV");
-		const btn_update = document.createElement("BUTTON");
-		btn_update.textContent = "Actualizar"
-		btn_update.addEventListener("click", function() {
-			sessionStorage.setItem("productoId", producto.id);
-			window.location = `${BASE_URL}app/owner/inventario/actualizar`;
-		})
-		divAcciones.appendChild(btn_update);
-		const btn_delete = document.createElement("BUTTON");
-		btn_delete.textContent = "Eliminar"		
-		btn_delete.addEventListener("click", async function() {
-		await fetch(`${bBASE_URL}api/owner/inventario/eliminar/${producto.id}`, {
+		mostrarInventory.appendChild(divContenido);
+		document.querySelector(`#btn-update-${producto.id}`).addEventListener("click", update);
+        document.querySelector(`#btn-cancel-${producto.id}`).addEventListener("click", deleteProduct);
+	})
+}
+
+function update(event) {
+	sessionStorage.setItem("productoId", event.target.name);
+	window.location = `${BASE_URL}app/owner/inventario/actualizar`;
+}
+
+function deleteProduct(event) {
+	Swal.fire({
+		title: '¿Estas seguro?',
+		text: "No se podra reevertir los cambios!!",
+		icon: 'warning',
+		showCancelButton: true,
+		confirmButtonColor: '#3085d6',
+		cancelButtonColor: '#d33',
+		confirmButtonText: 'Eliminar'
+	}).then(async (result) => {
+		if (result.isConfirmed) {
+			const resultado = await fetch(`${BASE_URL}api/owner/inventario/eliminar/${event.target.name}`, {
 				method: 'DELETE',
 				headers: {
 						"Authorization": `Bearer ${sessionStorage.getItem("token")}`
@@ -79,14 +94,11 @@ function mostrarInventario(inventario) {
 			})
 			const respuesta = await resultado.json();
 			if(respuesta.errorCode == 404) {
-				let alertas = [respuesta.message];
-				mostrarAlerta(alertas);
+				alerta("Error", respuesta.message)
+				return;
 			} else {
 				document.location = `${BASE_URL}app/owner/inventario`;		
 			}
-		})
-		divAcciones.appendChild(btn_delete);
-		divContenido.appendChild(divAcciones);
-		mostrarInventory.appendChild(divContenido);
+		}
 	})
 }
